+++ CHANGELOG.md
@@ -42,6 +42,7 @@ Please view this file on the master branch, on stable branches it's out of date.
   - Modify GitHub importer to be retryable !7003
   - Fix and improve `Sortable.highest_label_priority`
   - Fixed sticky merge request tabs when sidebar is pinned
+  - Fix and improve `Sortable.highest_label_priority`
 
 ## 8.13.1 (2016-10-25)